requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==2.0.0
ecdsa==0.17.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
pyasn1==0.4.8
PyJWT==2.3.0
python-jose==3.3.0
pytz==2021.3
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
